+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Include non full qualified vars on unused-public-var exclude filter.
 - Improve hover documentation: use correct markdown for docstrings; remove unnecessary new lines; add link to filename location.
+- Rollback full text changes on last release and change approach for a temporary fix. #424
 
 ## 2021.05.06-19.44.00